// node/configure/type/madoka-official-news.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type NoName1 = string;
export type NoName2 = string;

export interface NoName {
  view: View;
}
export interface View {
  init: NoName1;
  month: NoName2;
}